#373 Template - Separate read / write repository

+++ template/backend/microservice/src/Core/Optivem.Template.Core.Application.Interface/Customers/Queries/ListCustomersQueryResponse.cs
@@ -7,7 +7,7 @@ namespace Optivem.Template.Core.Application.Customers.Queries
     {
         public List<ListCustomersRecordResponse> Records { get; set; }
 
-        public int TotalRecords { get; set; }
+        public long TotalRecords { get; set; }
     }
 
     public class ListCustomersRecordResponse
